bug fix for skip

+++ leveldb/iterator.go
@@ -137,7 +137,7 @@ func (it *Iterator) Next() {
 }
 
 func (it *Iterator) Skip(offset int64) {
-	for i := 0; i < offset; i++ {
+	for i := int64(0); i < offset; i++ {
 		if it.Valid() {
 			it.Next()
 		}
